%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{fitCalMaTe}
\alias{refineCN}
\alias{truncateThetaAB}
\alias{truncateThetaAB.array}
\alias{CalMaTeNormalization}
\alias{allocateOutputDataSets}
\alias{allocateOutputDataSets.CalMaTeNormalization}
\alias{as.character.CalMaTeNormalization}
\alias{getAsteriskTags}
\alias{getAsteriskTags.CalMaTeNormalization}
\alias{getDataSets}
\alias{getDataSets.CalMaTeNormalization}
\alias{getFullName}
\alias{getFullName.CalMaTeNormalization}
\alias{getInputDataSet}
\alias{getInputDataSet.CalMaTeNormalization}
\alias{getName.CalMaTeNormalization}
\alias{getOutputDataSet}
\alias{getOutputDataSet.CalMaTeNormalization}
\alias{getPath.CalMaTeNormalization}
\alias{getRootPath}
\alias{getRootPath.CalMaTeNormalization}
\alias{getTags}
\alias{getTags.CalMaTeNormalization}
\alias{nbrOfFiles}
\alias{nbrOfFiles.CalMaTeNormalization}
\alias{process}
\alias{process.CalMaTeNormalization}
\alias{setTags}
\alias{setTags.CalMaTeNormalization}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}


\keyword{documentation}
\keyword{internal}
