%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{calmate-package}
\alias{calmate-package}
\alias{calmate}
\docType{package}

\title{Package calmate}


 \description{
   NNN.

   This package should be considered to be in an alpha phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires...
 }

 \section{Installation and updates}{
   TBA.
 }

 \section{To get started}{
   TBA.
 }

 \author{
   Angel Rubio <arubio@ceit.es>,
   Maria Ortiz <mortiz@ceit.es>,
   Henrik Bengtsson <henrikb@braju.com>
 }

 \section{License}{
   TBA.
 }

 \references{
   TBA.
 }
\keyword{package}
