%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{calmate-package}
\alias{calmate-package}
\alias{calmate}
\docType{package}

\title{Package calmate}


 \description{
   Post-calibration of allele-specific copy-number estimates (ASCNs) [UNDER HEAVY DEVELOPLMENT].

   This package should be considered to be in an alpha phase.
   You should expect the API to be changing over time.
 }

 \section{Requirements}{
   This package requires...
 }

 \section{Installation and updates}{
   TBA.
 }

 \section{To get started}{
  \enumerate{
   \item To process SNP and non-polymorphic signals, see \code{\link{calmateByTotalAndFracB}}().  If you are working solely with SNP signals, \code{\link{calmateByThetaAB}}() is also available, but we recommend the former.
   \item For processing data in the aroma framework, see \code{\link{CalMaTeCalibration}}.
  }
 }

 \author{
   Maria Ortiz-Estevez <mortiz@ceit.es>,
   Ander Aramburu,
   Henrik Bengtsson <henrikb@braju.com>,
   Pierre Neuvial <pierre@stat.berkeley.edu>,
   Angel Rubio <arubio@ceit.es>.
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.
 }

 \references{
   TBA.
 }
\keyword{package}
