%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{truncateThetaAB}
\alias{truncateThetaAB.array}
\alias{allocateOutputDataSets}
\alias{allocateOutputDataSets.CalMaTeCalibration}
\alias{as.character.CalMaTeCalibration}
\alias{getAsteriskTags}
\alias{getAsteriskTags.CalMaTeCalibration}
\alias{getDataSets}
\alias{getDataSets.CalMaTeCalibration}
\alias{getFullName}
\alias{getFullName.CalMaTeCalibration}
\alias{getInputDataSet}
\alias{getInputDataSet.CalMaTeCalibration}
\alias{getName.CalMaTeCalibration}
\alias{getOutputDataSet}
\alias{getOutputDataSet.CalMaTeCalibration}
\alias{getPath.CalMaTeCalibration}
\alias{getRootPath}
\alias{getRootPath.CalMaTeCalibration}
\alias{getTags}
\alias{getTags.CalMaTeCalibration}
\alias{nbrOfFiles}
\alias{nbrOfFiles.CalMaTeCalibration}
\alias{process}
\alias{process.CalMaTeCalibration}
\alias{setTags}
\alias{setTags.CalMaTeCalibration}
\alias{findUnitsTodo}
\alias{findUnitsTodo.CalMaTeCalibration}
\alias{getOutputDataSets}
\alias{getOutputDataSets.CalMaTeCalibration}
\alias{truncateFracB}
\alias{truncateFracB.array}
\alias{truncateFracB.matrix}
\alias{truncateThetaAB.matrix}
\alias{getReferences}
\alias{getReferences.CalMaTeCalibration}
\alias{cat}
\alias{sapply}
\alias{getParameters.CalMaTeCalibration}
\alias{getParametersAsString.CalMaTeCalibration}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}


\keyword{documentation}
\keyword{internal}
